tpl_modules_attributes.php potentially displays extra option names and data Issue #1
An issue was reported on a site that had tpl_modules_attributes.php installed as written where the product had only 2 option names, but three would be displayed onscreen with the third duplicating one of the previous option names and options.  The first two would have the dynamic dropdowns applied.  This is thought to be possibly related to the use of $inSBA as an object and that the object was modified elsewhere in the code before final execution in this section.  As a result a separate variable was instantiated to potentially alleviate the cause.  If this is successful, then perhaps the code could/would be restored, but to also perform the evaluation again just prior to where the result is specifically needed.